requirements3.11.txt
beautifulsoup4==4.12.2
certifi==2021.10.8
charset-normalizer==2.0.8
colorama==0.4.4
google==3.0.0
googlesearch-python==1.2.3
graphviz==0.19
idna==3.3
jsonpath-ng==1.6.0
ply==3.11
python-dotenv==1.0.0
requests==2.26.0
scapy==2.5.0
soupsieve==2.5
tqdm==4.66.1
urllib3==1.26.7
zoomeye==2.2.0
